# lib/api-zod/src/generated/types/generateQuestionsBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateQuestionsBody {
  /** MP History or MP Geography */
  subject: string;
  topic?: string;
  /** Number of questions to generate (max 20) */
  count: number;
  focusWeakTopics?: boolean;
}