# lib/api-zod/src/generated/types/question.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Question {
  id: number;
  questionText: string;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  correctOption: string;
  explanation: string;
  subject: string;
  topic: string;
  difficulty: string;
  timesAnswered: number;
  timesWrong: number;
  createdAt: Date;
}