# lib/api-zod/src/generated/types/sessionResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Session } from "./session";

export interface SessionResult {
  session: Session;
  wrongTopics: string[];
  correctCount: number;
  wrongCount: number;
  score: number;
}