# lib/api-zod/src/generated/types/statsOverview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StatsOverview {
  totalQuestions: number;
  totalSessions: number;
  averageScore: number;
  totalCorrect: number;
  totalWrong: number;
  weakTopicsCount: number;
  streak: number;
}