# lib/api-zod/src/generated/types/topicStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TopicStat {
  topic: string;
  subject: string;
  totalAttempts: number;
  correctCount: number;
  accuracy: number;
}